remove strategy
