+++ src/hipp/kh9pc/__init__.py
@@ -2,26 +2,6 @@ from . import image_mosaic
 from . import quality_control as qc
 from .batch import join_images
 
-from .restitution import (
-    CollimationRectificationStrategy,
-    FlatRectificationStrategy,
-    ImageTransformer,
-    ImageTransformerAffine,
-    ImageTransformerTps,
-    PolyRectificationStrategy,
-    RectificationStrategy,
-)
+from .restitution import control_points, detectors, output_size, plotters
 
-__all__ = [
-    "image_mosaic",
-    "ImageTransformer",
-    "ImageTransformerTps",
-    "ImageTransformerAffine",
-    "RectificationStrategy",
-    "FlatRectificationStrategy",
-    "PolyRectificationStrategy",
-    "CollimationRectificationStrategy",
-    # legacy — still used by core.py
-    "join_images",
-    "qc",
-]
+__all__ = ["image_mosaic", "join_images", "qc", "control_points", "detectors", "output_size", "plotters"]
